move code to TaskList component
